diesel-orm-dev 完成diesel orm持久层切换

+++ src/utils/jwt_util.rs
@@ -1,12 +1,14 @@
-use jsonwebtoken::{decode, encode, DecodingKey, EncodingKey, Header, Validation, errors::ErrorKind};
-use serde::{Deserialize, Serialize};
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
+
+use jsonwebtoken::{decode, DecodingKey, encode, EncodingKey, errors::ErrorKind, Header, Validation};
+use serde::{Deserialize, Serialize};
+
 use crate::utils::error::WhoUnfollowedError;
-use crate::utils::error::WhoUnfollowedError::{JwtTokenError};
+use crate::utils::error::WhoUnfollowedError::JwtTokenError;
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct JWTToken {
-    pub id: i32,
+    pub id: i64,
     pub username: String,
     pub permissions: Vec<String>,
     aud: String,
@@ -24,7 +26,7 @@ pub struct JWTToken {
 }
 
 impl JWTToken {
-    pub fn new(id: i32, username: &str, permissions: Vec<String>) -> JWTToken {
+    pub fn new(id: i64, username: &str, permissions: Vec<String>) -> JWTToken {
         let now = SystemTime::now();
         //过期时间
         let m30 = Duration::from_secs(1800000);
